Upon initialisation, read the XML file
Current BeID development tools work with XML files to represent the data
on the card. Reuse that format, so that existing developers don't have
to regenerate their cards (and so that a certain level of
interoperability between those tools exists)

We may add another option in the future, but that's not for now.

Also, add the Belpic applet's AID as the DF name for the MF