Make the Relay card still parse and log the APDU
The basic VirtualSmartcard implementation parses and logs the APDU sent
to the card in the execute instance method, and logs the received APDU
after calling that method.

The Relay implementation, however, overrides the execute method, but
does not log it. The result, when running at the info log level, is that
you see the reply from the card, but not the request from the
application. This is confusing.

Additionally, if the relay implementation shows the request APDU, then
it can be used as a method to analyze what an application is trying to
do.

To remedy all that, parse the APDU and log that parsed value, but don't
do anything further with it.

Signed-off-by: Wouter Verhelst <w@uter.be>