Filter the SELECT FILE command
The spec for the Belgian electronic ID card interprets the ISO 7816
"SELECT FILE" command in a somewhat simplified way. First, the MF (3F00)
can be selected from anywhere when P1 = 0x02; second, child DFs are
selected also with P1 = 0x02. The applet does not interpret or treat
specially cases where P1 = 0x00 or P1 = 0x01.

In order to not to have to rewrite the "select file" logic that already
exists in the Iso7816OS class, filter those two commands so that they
contain the P1 value which the Iso7816OS class would search for.