Filter out warnings
The belpic applet does not deal with file sizes. It just returns less
data than was requested, and the remote is supposed to figure out that
the file was smaller than it assumed before based on that. It still
returns SW12 0x9000 in that case, however.

Add a "formatResult" method which filters out the SW12 = 0x6282 case, so
that applications don't see values they don't expect.